# lib/api-zod/src/generated/types/aiMemory.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * OrderOS order intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { AiMemoryState } from './aiMemoryState';

export interface AiMemory {
  id: string;
  dealerId: string;
  dealerName: string;
  trigger: string;
  interpretation: string;
  confidence: number;
  source: string;
  createdAt: string;
  lastUsedAt: string;
  usageCount: number;
  state: AiMemoryState;
}